pyproject.toml
[project]
name = "goit-py-web-hw-11"
version = "0.1.0"
description = ""
authors = [
    {name = "Oleksii Haiduk",email = "Silvestr171@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic (>=1.16.5,<2.0.0)",
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
